feat(processor): move the line item logic to the capture.converter
